# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manga Sekai Project Backend",
        "description": "Backend for accessing and managing the Manga data hosted in a Manga Sekai Project instance",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/auth": {
            "summary": "Authentication"
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Performs login",
                "description": "Login to the application",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token information after successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Performs logout",
                "description": "Logout from the application",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Successful logout"
                    }
                },
                "security": [
                    {
                        "Token": []
                    }
                ]
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Refreshes a valid token",
                "description": "Refreshes a valid token",
                "operationId": "refreshToken",
                "responses": {
                    "200": {
                        "description": "Token information after successful refresh",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TokenResponse": {
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "token_type": {
                        "type": "string"
                    },
                    "expires_in": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Token": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "auth",
            "description": "Authentication"
        }
    ]
}